document / cleanup
